+++ src/main/java/ua/com/apricotka/repository/OrderItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ua.com.apricotka.entity.OrderItem;
 
 @Repository
-public interface OrderItemRepository extends CrudRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends CrudRepository<OrderItem, Long> {
 }
